+++ src/main/java/baseball/model/Ball.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.model;
 
 import java.util.Objects;
