+++ src/test/java/baseball/BallTest.java
@@ -2,6 +2,7 @@ package baseball;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.model.Ball;
 import org.junit.jupiter.api.Test;
 
 public class BallTest {
