+++ src/test/java/baseball/NumberGenerateTest.java
@@ -2,6 +2,7 @@ package baseball;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.model.NumberGenerator;
 import org.junit.jupiter.api.Test;
 
 public class NumberGenerateTest {
