+++ src/test/java/baseball/NumberValidatorTest.java
@@ -3,6 +3,7 @@ package baseball;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import baseball.model.NumberValidator;
 import java.util.List;
 import org.junit.jupiter.api.Test;
